Configured node backend with: - linting (ESLint) - prettifying/formatting (Prettier) - git hooks (Husky, only hook is a format on staging) - Typescript (compiled files go to the ./build directory)
There are a number of scripts added to `package.json`, the majority of
which are cribbed from `https://github.com/alan2207/bulletproof-react`.
The notable ones are:
- `dev`: runs the code with a file watcher, auto-recompiles the code on
change
- `start`: validates the code (checks types/format/lint), builds
(compiles the JS files), and runs the compiled Javascript. Intended to
be used when running in production.
- `validate-and-build`: runs all the validation checks and builds if
successful

Also slightly changes the project structure, renamed the top-level `src`
folder to `pkg` as it contains packages (for `ui` and `backend`) which
themselves contain `src` folders (to house their own code).